refactoring things into WidgetLibrary
