+++ KlockWork/Views/Shared/WidgetLibrary/WidgetLibrary.UI.Sidebar.swift
@@ -13,6 +13,7 @@ import EventKit
 extension WidgetLibrary.UI {
     struct Sidebar {
         struct EventsWidget: View {
+            typealias Event = WidgetLibrary.UI.Individual.Event
             @EnvironmentObject public var state: Navigation
             @EnvironmentObject public var updater: ViewUpdater
             @StateObject public var ce: CoreDataCalendarEvent = CoreDataCalendarEvent(moc: PersistenceController.shared.container.viewContext)
@@ -24,7 +25,7 @@ extension WidgetLibrary.UI {
                 VStack(alignment: .leading, spacing: 1) {
                     if self.calendar > 0 {
                         ForEach(self.upcomingEvents.sorted(by: {$0.startDate < $1.startDate}), id: \.self) { event in
-                            SingleEvent(event: event)
+                            Event(event: event)
                         }
                     } else {
                         HStack(alignment: .center, spacing: 0) {
@@ -42,35 +43,6 @@ extension WidgetLibrary.UI {
                 .id(updater.get("dashboard.header"))
             }
         }
-
-        struct SingleEvent: View {
-            @EnvironmentObject public var state: Navigation
-            public let event: EKEvent
-            @State private var hasEventPassed: Bool = false
-
-            var body: some View {
-                SidebarItem(
-                    data: "\(event.startTime()) - \(event.endTime()): \(event.title ?? "")",
-                    help: "\(event.startTime()) - \(event.endTime()): \(event.title ?? "")",
-                    icon: "chevron.right",
-                    orientation: .right,
-                    action: {},
-                    showBorder: false,
-                    showButton: false
-                )
-                .background(self.hasEventPassed ? Theme.lightWhite : .orange)
-                .foregroundStyle(self.hasEventPassed ? Theme.lightBase : Theme.base)
-                .onAppear(perform: self.actionOnAppear)
-            }
-        }
-    }
-}
-
-extension WidgetLibrary.UI.Sidebar.SingleEvent {
-    /// Onload handler. Determines if event has passed or not
-    /// - Returns: Void
-    private func actionOnAppear() -> Void {
-        self.hasEventPassed = event.startDate < Date()
     }
 }
